cambios pedientes de actualizar al 20200401

+++ assets/js/index.js
@@ -1,4 +1,4 @@
-$(document).ready(function() {
+$(document).ready(function () {
   ShowTotal()
   // Llamado a la lista de productos
   getProducts()
@@ -60,9 +60,9 @@ const createProductTemplate = (name, price, stock, category, imageSrc, i) => {
     <p class="product__stock card-text"><strong>Stock:</strong> ${stock}</p>
     <p class="card-text"><strong>Categoría:</strong>  ${i} <a href="#">${category}</a></p>
 
-    <p class="card-text"><strong>Cantidad:</strong> <input type="number" class="id-${i}" value="1" id="cantProducts">
+    <p class="card-text"><strong>Cantidad:</strong> <input type="number" class="id-${i} inputCAntidad" value="1" id="cantProducts">
 
-    <button class="btn btn-success" onclick="addProduct('${name}','${imageSrc}','${price}','${stock}','${category}', $('.id-${i}').val())">Agregar al carro</button>
+    <button class="btn btn-success botonAgregar" onclick="addProduct('${name}','${imageSrc}','${price}','${stock}','${category}', $('.id-${i}').val())">Agregar al carro</button>
   </div>
   
 `
@@ -85,21 +85,21 @@ const createPaginationTemplate = pages => {
 // Lógica para la busqueda general
 //falta entonces que navegue entre las paginas
 
-$('#search').keyup(function(){
+$('#search').keyup(function () {
   let productoTitle = $('.card-title');
   let buscando = $(this).val();
-  let item='';
-  for( let i = 0; i < productoTitle.length; i++ ){
-      item = $(productoTitle[i]).html().toLowerCase();
-       for(let x = 0; x < item.length; x++ ){
-           if( buscando.length == 0 || item.indexOf( buscando ) > -1 ){
-               $(productoTitle[i]).parents('.card-title').show(); 
-           }else{
-                $(productoTitle[i]).parents('.card-title').hide();
-           }
-       }
-  }  
-  })
+  let item = '';
+  for (let i = 0; i < productoTitle.length; i++) {
+    item = $(productoTitle[i]).html().toLowerCase();
+    for (let x = 0; x < item.length; x++) {
+      if (buscando.length == 0 || item.indexOf(buscando) > -1) {
+        $(productoTitle[i]).parents('.card-title').show();
+      } else {
+        $(productoTitle[i]).parents('.card-title').hide();
+      }
+    }
+  }
+})
 
 // Fin de la logica para la busqueda general
 
@@ -114,50 +114,56 @@ function addProduct(name, imageSrc, price, stock, category, cantProduct) {
       })
 
       localStorage.setItem('carrito', JSON.stringify(detalleAlCarro))
+      alert('Producto Guardado en el carrito')
     } else {
       detalleAlCarro = JSON.parse(localStorage.getItem('carrito'))
-      detalleAlCarro.push({
-        data: { name, price, stock, imageSrc, category, cantProduct: cantProduct },
-      })
-      localStorage.setItem('carrito', JSON.stringify(detalleAlCarro))
+      verificarProdYaSelected(name, imageSrc, price, stock, category, cantProduct, detalleAlCarro);
     }
-
-    alert('Producto Guardado en el carrito')
   } else {
     alert('Ingrese una cantidad lógica')
   }
-
-  ShowTotal()
+  ShowTotal();
 }
 
-// function verificarProdYaSelected(name, imageSrc, price, stock, category, cantProduct, detalleAlCarro){
-//   detalleAlCarro.forEach(element => {
-//     if((name + imageSrc + price + stock + category) == (element.data.name + element.data.imageSrc + element.data.price + element.data.stock + element.data.category)){
-//       Swal.fire({
-//         showClass: {
-//           popup: 'animated fadeInDown faster'
-//         },
-//         hideClass: {
-//           popup: 'animated fadeOutUp faster'
-//         },
-//         title: 'Aviso, contenido del carrito',
-//         text: `El producto seleccionado "${name}" ya tiene ${element.data.cantProduct} unidades en el carrito.`,
-//         icon: 'warning',
-//         showCancelButton: true,
-//         cancelButtonText: 'Cancelar',
-//         confirmButtonColor: '#3085d6',
-//         cancelButtonColor: '#d33',
-//         confirmButtonText: 'Agregar'
-//       }).then((result) => {
-//         if (result.value) {
-//           element.data.cantProduct = parseInt(element.data.cantProduct) + parseInt(cantProduct);
-//           localStorage.setItem('carrito', JSON.stringify(detalleAlCarro))
-//           alert('Producto Guardado en el carrito')
-//         }
-
-
-
-
+function verificarProdYaSelected(name, imageSrc, price, stock, category, cantProduct, detalleAlCarro) {
+  let prodYaExiste = false;
+  detalleAlCarro.forEach(element => {
+    if ((name + imageSrc + price + stock + category) == (element.data.name + element.data.imageSrc + element.data.price + element.data.stock + element.data.category)) {
+      prodYaExiste = true;
+      Swal.fire({
+        showClass: {
+          popup: 'animated fadeInDown faster'
+        },
+        hideClass: {
+          popup: 'animated fadeOutUp faster'
+        },
+        title: 'Aviso, contenido del carrito',
+        text: `El producto seleccionado "${name}" ya tiene ${element.data.cantProduct} unidades en el carrito. ¿Quiere agregar ${cantProduct} más?`,
+        icon: 'warning',
+        showCancelButton: true,
+        cancelButtonText: 'Cancelar',
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Agregar'
+      }).then((result) => {
+        if (result.value) {
+          element.data.cantProduct = parseInt(element.data.cantProduct) + parseInt(cantProduct);
+          localStorage.setItem('carrito', JSON.stringify(detalleAlCarro))
+          alert('Producto Guardado en el carrito')
+          ShowTotal();
+        }
+      })
+    }
+  });
+  if (!prodYaExiste) {
+    detalleAlCarro.push({
+      data: { name, price, stock, imageSrc, category, cantProduct: cantProduct }
+    })
+    localStorage.setItem('carrito', JSON.stringify(detalleAlCarro))
+    alert('Producto Guardado en el carrito')
+    ShowTotal();
+  }
+}
 
 let page = 0
 
@@ -193,12 +199,14 @@ function atrasPage() {
 }
 
 const ShowTotal = () => {
-  let localsito = JSON.parse(localStorage.getItem('carrito'))
+  if (localStorage.carrito) {
+    let localsito = JSON.parse(localStorage.getItem('carrito'))
 
-  let total = 0
-  localsito.forEach(p => {
-    total += parseInt(p.data.price) * parseInt(p.data.cantProduct)
-  })
+    let total = 0
+    localsito.forEach(p => {
+      total += parseInt(p.data.price) * parseInt(p.data.cantProduct)
+    })
 
-  $('#total').html(total.toString().replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1.'))
+    $('#total').html(total.toString().replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1.'))
+  }
 }
